Tv schedule (#2544)
* sql

* schema

* basic schedule

* stream title

* separate tv pages

* modify, delete events

* is_featured, schedule page, admin permission check

* confirm delete

* add hour buffer, styling

* refactor

* seo title

* url processing

* improve schedule form

* word

* sidebar

* twitch support

* useTVSchedule

* remove debug

* fix unrelated timeline error

+++ backend/api/src/set-tv.ts
@@ -1,45 +1,77 @@
-import { APIError, authEndpoint, validate } from './helpers/endpoint'
+import { authEndpoint, validate } from './helpers/endpoint'
 import { z } from 'zod'
 import {
   createSupabaseClient,
   createSupabaseDirectClient,
 } from 'shared/supabase/init'
 import { getContractIdFromSlug } from 'shared/supabase/contracts'
-import { isAdminId, isModId } from 'common/envs/constants'
+import { isAdminId } from 'common/envs/constants'
 
-const schema = z
-  .object({
-    slug: z.string(),
-    streamId: z.string(),
-    source: z.string(),
-  })
-  .optional()
+const schema = z.object({
+  id: z.string().optional(),
+  slug: z.string(),
+  streamId: z.string(),
+  source: z.union([z.literal('youtube'), z.literal('twitch')]),
+  title: z.string(),
+  startTime: z.string(),
+  endTime: z.string(),
+  isFeatured: z.boolean(),
+})
 
-export const settv = authEndpoint(async (req, auth, log) => {
-  const isMod = isAdminId(auth.uid) || isModId(auth.uid)
-  if (!isMod) {
-    throw new APIError(403, 'You are not authorized to set tv settings')
-  }
-
+export const settv = authEndpoint(async (req, auth) => {
   const tvSettings = validate(schema, req.body)
 
   const pg = createSupabaseDirectClient()
 
-  if (!tvSettings) {
-    await pg.none('delete from tv_schedule where id = 1')
-    return { status: 'success' }
-  }
+  const { id, slug, streamId, source, title, startTime, endTime } = tvSettings
 
-  const { slug, streamId, source } = tvSettings
-  const processedStreamId = streamId.trim().substring(0, 11)
+  const isFeatured = isAdminId(auth.uid) && tvSettings.isFeatured
 
   const db = createSupabaseClient()
   const contractId = await getContractIdFromSlug(db, slug)
 
+  if (id) {
+    await pg.none(
+      'UPDATE tv_schedule SET contract_id = $1, stream_id = $2, source = $3, title = $4, start_time = $5, end_time = $6, is_featured = $7 WHERE id = $8',
+      [contractId, streamId, source, title, startTime, endTime, isFeatured, id]
+    )
+    return { status: 'success' }
+  }
+
   await pg.none(
-    'update tv_schedule set contract_id = $1, stream_id = $2, source = $3 where id = 1',
-    [contractId, processedStreamId, source]
+    'INSERT INTO tv_schedule (creator_id, contract_id, stream_id, source, title, start_time, end_time, is_featured) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)',
+    [
+      auth.uid,
+      contractId,
+      streamId,
+      source,
+      title,
+      startTime,
+      endTime,
+      isFeatured,
+    ]
   )
 
+  return { status: 'success' }
+})
+
+const deleteSchema = z.object({
+  id: z.string(),
+})
+
+export const deletetv = authEndpoint(async (req, auth) => {
+  const { id } = validate(deleteSchema, req.body)
+  const userId = auth.uid
+
+  const pg = createSupabaseDirectClient()
+
+  if (isAdminId(userId)) {
+    await pg.none('delete from tv_schedule where id = $1', [id])
+  } else {
+    await pg.none('delete from tv_schedule where id = $1 and creator_id = $2', [
+      id,
+      userId,
+    ])
+  }
   return { status: 'success' }
 })
